refactor: 삭제 메서드 타입 변환

+++ src/main/java/kr/bi/greenmate/repository/CommunityPostCommentRepository.java
@@ -22,5 +22,5 @@ public interface CommunityPostCommentRepository extends JpaRepository<CommunityP
 	@EntityGraph(attributePaths = "user")
 	List<CommunityPostComment> findByParent_IdAndIdLessThanOrderByIdDesc(Long postId, Long lastId, Pageable pageable);
 
-	int deleteByIdAndParent_IdAndUser_Id(Long commentId, Long postId, Long userId);
+	long deleteByIdAndParent_IdAndUser_Id(Long commentId, Long postId, Long userId);
 }
